api/tut/requirements.txt
django
djangorestframework
pyyaml
requests
django-cors-headers

# pip install -r requirement.txt